fix(ci): batch -Werror cleanups for pythonpkg native/numpy/pandas/scan/wrapper
Sweep up the remaining GCC 14 -Werror diagnostics from the manylinux2014
build of integration/python (old-style-cast, sign-conversion, sign-promo,
sign-compare, reorder, unused-parameter, unused-variable, unused-value,
extra/copy-ctor base init).

In python_replacement_scan FetchObjectData also replace the stub
result_chunk.append(...) (no such method on data_chunk_t) with an explicit
chunk collection + row-by-row merge, so multi-batch pandas/numpy scans
build a single accumulated data_chunk for make_node_raw_data.

+++ integration/python/numpy/type.cpp
@@ -177,7 +177,7 @@ components::types::complex_logical_type NumpyToLogicalType(const NumpyType &col_
 		return logical_type::TIMESTAMP_US;
 	}
 	default:
-		throw std::runtime_error("No known conversion for NumpyNullableType "+to_string(uint8_t(col_type.type))+" to logical_type");
+		throw std::runtime_error("No known conversion for NumpyNullableType "+to_string(static_cast<unsigned int>(col_type.type))+" to logical_type");
 	}
 }
